requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
boto3
botocore
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Deprecated==1.2.14
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.4.0
django-user-agents==0.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
Faker==28.0.0
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
gunicorn==21.2.0
idna==3.6
importlib_resources==6.4.0
jmespath==1.0.1
kiwisolver==1.4.5
line-bot-sdk==3.9.0
matplotlib==3.9.0
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
ua-parser==0.18.0
urllib3
user-agents==2.2.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.1
whitenoise
psycopg2-binary
dj-database-url
